Fix: channel페이지 팔로우 수정, 조회수 문제 해결 - 조회수 2씩 오르는 문제 해결(cardDetailContent컴포넌트를 CardDetailPage로 병합) - 팔로우 여부 불러오는 api 수정 - channel페이지에 Follow 컴포넌트 추가 - channel페이지에서 팔로우 기능 수정완료
