--- src/components/CardDetailContent/CardDetailContent.scss
@@ -1,289 +1,0 @@
-.detail-out-wrapper {
-  width: 100%;
-  margin: 0 auto;
-  .detail-header-wrapper {
-    width: 896px;
-    margin: 38px auto;
-    max-width: 100%;
-    .detail-title-wrapper {
-      .detail-title {
-        margin-bottom: 0;
-      }
-    }
-    .detail-writer-by {
-      color: #a1a1a1;
-      padding-right: 5px;
-      font-style: italic;
-      font-size: 14px;
-    }
-    .detail-writer-nickname {
-      padding-right: 10px;
-      border-right: 2px solid #dddddd;
-      font-size: 14px;
-      background-color: #ffffff;
-      cursor: pointer;
-    }
-    .detail-date {
-      padding: 0px 10px;
-      border-right: 2px solid #dddddd;
-      color: #a1a1a1;
-      font-size: 14px;
-    }
-    .detail-inquiry-count {
-      padding: 0px 10px;
-      color: #a1a1a1;
-      font-size: 14px;
-      margin-right: 625px;
-    }
-    .menuIcon {
-      display: inline-block;
-      background: url('./../../assets/image/menu.png');
-      background-repeat: no-repeat;
-      background-size: 23px;
-      border: none;
-      width: 23px;
-      height: 23px;
-      cursor: pointer;
-    }
-    .menuSelectWrapper {
-      float: left;
-      border: 1px solid #ddd;
-      background-color: #fff;
-      width: 152px;
-      margin-left: 745px;
-      position: relative;
-      // z-index: 101;
-      .selectModify {
-        height: 41px;
-        padding: 10px 10px 10px 13px;
-        font-size: 15px;
-        border-bottom: 1px solid rgb(241, 241, 241);
-        // z-index: 100;
-        cursor: pointer;
-        .selectPen {
-          display: inline-block;
-          background: url('./../../assets/image/pen.png');
-          background-repeat: no-repeat;
-          background-size: 15px;
-          border: none;
-          width: 15px;
-          height: 15px;
-          margin-left: 45px;
-          // z-index: 100;
-        }
-      }
-      .selectDelete {
-        height: 41px;
-        padding: 10px 10px 10px 13px;
-        font-size: 15px;
-        cursor: pointer;
-        // z-index: 100;
-        .selectBin {
-          display: inline-block;
-          background: url('./../../assets/image/bin.png');
-          background-repeat: no-repeat;
-          background-size: 15px;
-          border: none;
-          width: 15px;
-          height: 15px;
-          margin-left: 45px;
-          // z-index: 100;
-        }
-      }
-    }
-  }
-
-  .detail-content-wrapper {
-    width: 896px;
-    margin: auto;
-    // z-index: 200;
-    .detail-content-pictures {
-      width: 100%;
-      img {
-        display: block;
-        width: 100%;
-        height: auto;
-        text-align: center;
-      }
-    }
-    .detail-content-content {
-      margin: 30px 0px;
-    }
-    .detail-tag-wrapper {
-      margin: 20px 0px 15px 0px;
-    }
-    .detail-copy-right {
-      margin: 0px 0px 150px 0px;
-      color: #a1a1a1;
-      font-size: 13px;
-    }
-  }
-
-  .detail-reaction-wrapper {
-    display: flex;
-    width: 100%;
-    justify-content: center;
-    margin-bottom: 2.5%;
-    .detail-reaction-inner-wrapper {
-      display: flex;
-      flex-direction: column;
-      .detail-reaction-icon-wrapper {
-        text-align: center;
-        .detail-icon {
-          background-color: white;
-          img {
-            width: 38px;
-            height: 38px;
-            cursor: pointer;
-            background-color: white;
-          }
-        }
-      }
-    }
-    .detail-reaction-inner-wrapper {
-      display: flex;
-      flex-direction: column;
-      .detail-reaction-icon-wrapper {
-        text-align: center;
-        .detail-icon {
-          background-color: white;
-          img {
-            width: 38px;
-            height: 38px;
-            cursor: pointer;
-            background-color: white;
-          }
-        }
-      }
-      .detail-reaction-icon-second-wrapper {
-        flex-direction: column;
-        width: 100px;
-        .detail-icon-title {
-          font-size: 12px;
-          text-align: center;
-          color: #aaaaaa;
-          width: 100%;
-        }
-        .detail-icon-count {
-          font-size: 16px;
-          text-align: center;
-        }
-      }
-    }
-  }
-
-  hr {
-    width: 896px;
-    border: 0;
-    height: 0.25px;
-    background: #ececec;
-  }
-
-  .detail-reply-out-wrapper {
-    position: relative;
-    width: 896px;
-    height: 30px;
-    margin: 0 auto;
-    margin-bottom: 20px;
-
-    .detail-reply-wrapper {
-      position: absolute;
-      width: 100%;
-      margin-left: -25px;
-      .detail-reply-reaction-icon {
-        vertical-align: middle;
-        img {
-          width: 25px;
-          height: 27px;
-        }
-      }
-      .detail-reply-reaction-count {
-        margin-left: 10px;
-      }
-      .detail-reply-icon {
-        vertical-align: middle;
-        margin-left: 30px;
-        img {
-          width: 25px;
-          height: 27px;
-        }
-      }
-      .detail-reply-count {
-        margin-left: 10px;
-        margin-right: 5px;
-      }
-    }
-  }
-  details {
-    width: 896px;
-    margin: 0 auto;
-    background-color: #ffffff;
-    position: relative;
-    cursor: pointer;
-    summary {
-      width: 300px;
-      text-align: left;
-      top: -20%;
-    }
-    .detail-reply-input-wrapper {
-      width: 896px;
-      height: 176px;
-      background-color: #fff;
-      border: 1px solid #b3b3b3;
-      margin: 0 auto;
-      border-radius: 3px;
-      .detail-reply-input-inner-wrapper {
-        padding: 10px;
-        .detail-reply-title {
-          width: 870px;
-          height: 23px;
-          font-size: 13px;
-          color: #000000;
-          font-weight: bold;
-        }
-        .detail-reply-text-area-wrapper {
-          border: none;
-          width: 870px;
-          height: 49px;
-          textarea {
-            border: none;
-            width: 100%;
-            height: 100%;
-            resize: none;
-            overflow: hidden;
-          }
-          textarea:focus {
-            outline: none;
-          }
-          textarea::placeholder {
-            font-size: 12px;
-            color: #cccccc;
-          }
-        }
-        .detail-reply-text-limit {
-          text-align: right;
-          height: 22px;
-          color: #999999;
-          font-size: 13px;
-          margin-top: 15px;
-        }
-      }
-      .detail-reply-apload-btn {
-        border-top: 1px solid #ededed;
-        text-align: right;
-        padding-bottom: 0;
-        button {
-          width: 68px;
-          height: 45px;
-          background-color: #00d084;
-          font-size: 13px;
-          font-weight: bold;
-          line-height: 45px;
-          color: #fff;
-          border-radius: 0px 0px 3px 0px;
-          cursor: pointer;
-        }
-      }
-    }
-  }
-}
